rpi4: Prepare for platmods
Restructure the `rpi4` directory to look very similar to
`sun4v/ontario`, which is the reference I'm using for platmods.

There are two meaningful changes here:
* What will become the `platmod` (`rpi4.c`) moves to `os/`.
* Drivers are simplified (the headers are unnecessary).
